Flatten transparency in gridlines
Currently, a plot with default settings has transparent gridlines,
which can slow down printing (issue #4940). This commit replaces
gridline colours with an opaque blend of the chosen and background
colours, such that the result looks the same but doesn't use
transparency.

The color blending formula is from the PDF specification (ISO
32000:2008, Section 11.3), and respects background alpha values so
that e.g. gridlines on a transparent background are still
semi-transparent.